+++ plugin.audio.podcasts/addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.audio.podcasts" name="RSS Podcasts" version="2.1.0" provider-name="Heckie">
+<addon id="plugin.audio.podcasts" name="RSS Podcasts" version="2.2.1" provider-name="Heckie">
 	<requires>
 		<import addon="xbmc.python" version="3.0.0" />
 		<import addon="script.module.requests" version="2.25.1" />
@@ -26,6 +26,18 @@ Dateien direkt eingehängt werden.
 		<website>https://github.com/Heckie75/kodi-addon-podcast</website>
 		<source>https://github.com/Heckie75/kodi-addon-podcast</source>
 		<news>
+v2.2.1 (2021-10-17)
+- less restrictive when parsing RSS feeds with no XML processing instruction (reported issue #16)
+- Forced UTF-8 encoding, reported problem starting with Kodi 19.2 (reported issue #15)
+
+v2.2.0 (2021-08-14)
+- Added feauture in order to backup feeds as OPML file
+
+v2.1.1 (2021-08-13)
+- Migration to new settings format
+- Refactoring only. No new features
+- Minor bugfixes
+
 v2.1.0 (2021-05-24)
 - Added support for importing subscriptions from your gPodder account or from local files 
 - Added information like duration and plot for video podcasts
