+++ plugin.audio.podcasts/resources/language/resource.language.de_de/strings.po
@@ -161,10 +161,6 @@ msgctxt "#32051"
 msgid "Anchor for most recent podcast (good for bookmarking)"
 msgstr "Anker für aktuelle Sendung (hilfreich zum Bookmarken)"
 
-msgctxt "#32052"
-msgid "most recent episode"
-msgstr "aktuelle Sendung"
-
 msgctxt "#32053"
 msgid "Untitled podcast from"
 msgstr "Namenloser Podcast von"
@@ -285,26 +281,46 @@ msgctxt "#32088"
 msgid "There are no feeds to select"
 msgstr "Es gibt derzeit keine Feeds zur Auswahl"
 
+msgctxt "#32089"
+msgid "Backup subscription as OPML"
+msgstr "Feeds als OPML Datei sichern"
+
 msgctxt "#32090"
-msgid "Connection Error"
-msgstr "Verbindungsfehler"
+msgid "Select folder for backup"
+msgstr "Wähle Verzeichnis für Sicherung"
 
 msgctxt "#32091"
-msgid "HTTP Method %s not supported"
-msgstr "HTTP Methode %s wird nicht unterstützt"
+msgid "Backup successful"
+msgstr "Sicherung erfolgreich"
 
 msgctxt "#32092"
-msgid "Request Exception: Please check URL and port or see logs for further details"
-msgstr "Request Fehler: Prüfe URL und Port. Für weitere Details prüfe die Log-Datei"
-
-msgctxt "#32093"
-msgid "Unexpected HTTP Status %i for %s"
-msgstr "Unerwarteter HTTP Status %i für %s"
-
-msgctxt "#32094"
-msgid "Unexpected content for podcast"
-msgstr "Unerwarteter Inhalt für Podcast"
+msgid "Backup failed"
+msgstr "Sicherung gescheitert"
 
 msgctxt "#32095"
 msgid "Authentification at gPodder failed. Missing session"
 msgstr "Anmeldung an gPodder fehlgeschlagen"
+
+msgctxt "#32101"
+msgid "most recent episode"
+msgstr "aktuelle Sendung"
+
+msgctxt "#32151"
+msgid "Connection Error"
+msgstr "Verbindungsfehler"
+
+msgctxt "#32152"
+msgid "HTTP Method %s not supported"
+msgstr "HTTP Methode %s wird nicht unterstützt"
+
+msgctxt "#32153"
+msgid "Request Exception: Please check URL and port or see logs for further details"
+msgstr "Request Fehler: Prüfe URL und Port. Für weitere Details prüfe die Log-Datei"
+
+msgctxt "#32154"
+msgid "Unexpected HTTP Status %i for %s"
+msgstr "Unerwarteter HTTP Status %i für %s"
+
+msgctxt "#32155"
+msgid "Unexpected content for podcast"
+msgstr "Unerwarteter Inhalt für Podcast"
