+++ plugin.audio.podcasts/resources/language/resource.language.en_gb/strings.po
@@ -161,10 +161,6 @@ msgctxt "#32051"
 msgid "Anchor for most recent podcast (good for bookmarking)"
 msgstr ""
 
-msgctxt "#32052"
-msgid "most recent episode"
-msgstr ""
-
 msgctxt "#32053"
 msgid "Untitled podcast from"
 msgstr ""
@@ -285,26 +281,46 @@ msgctxt "#32088"
 msgid "There are no feeds to select"
 msgstr ""
 
+msgctxt "#32089"
+msgid "Backup subscription as OPML"
+msgstr ""
+
 msgctxt "#32090"
-msgid "Connection Error"
+msgid "Select folder for backup"
 msgstr ""
 
 msgctxt "#32091"
-msgid "HTTP Method %s not supported"
+msgid "Backup successful"
 msgstr ""
 
 msgctxt "#32092"
-msgid "Request Exception: Please check URL and port or see logs for further details"
-msgstr ""
-
-msgctxt "#32093"
-msgid "Unexpected HTTP Status %i for %s"
-msgstr ""
-
-msgctxt "#32094"
-msgid "Unexpected content for podcast"
+msgid "Backup failed"
 msgstr ""
 
 msgctxt "#32095"
 msgid "Authentification at gPodder failed. Missing session"
+msgstr ""
+
+msgctxt "#32101"
+msgid "most recent episode"
+msgstr ""
+
+msgctxt "#32151"
+msgid "Connection Error"
+msgstr ""
+
+msgctxt "#32152"
+msgid "HTTP Method %s not supported"
+msgstr ""
+
+msgctxt "#32153"
+msgid "Request Exception: Please check URL and port or see logs for further details"
+msgstr ""
+
+msgctxt "#32154"
+msgid "Unexpected HTTP Status %i for %s"
+msgstr ""
+
+msgctxt "#32155"
+msgid "Unexpected content for podcast"
 msgstr ""
